backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["instadeck"]

[project]
name = "instadeck-backend"
version = "0.1.0"
description = "InstaDeck API — document to outline to PPTX"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "python-pptx>=1.0.2",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "openai>=1.55.0",
    "Pillow>=11.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["instadeck", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"instadeck/parser_service.py" = ["E402"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
